% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Compare Clusters}
\usage{
compare_clusters(df, k, max.iter = 100, tol = 1e-04, scale = TRUE)
}
\arguments{
\item{df}{A dataframe containing the data to be clustered.}

\item{k}{A vector of integers representing the desired number of clusters (must be > 1).}

\item{max.iter}{An integer for the maximum number of iterations for convergence.}

\item{tol}{A numeric value specifying the tolerance for convergence.}

\item{scale}{Logical. If TRUE, scales the data before clustering; if FALSE, does not scale.}
}
\value{
A dataframe containing the following columns:
\item{distance}{The distance metric used.}
\item{k}{The number of clusters.}
\item{time}{Time taken for the clustering process (in seconds).}
\item{convergence}{The number of iterations to reach convergence.}
\item{DUNN}{The Dunn index for the clustering result.}
}
\description{
This function compares clustering results for different distance metrics and values of `k`.
}
\examples{

data <- data.frame(x = c(rnorm(20), rnorm(20,10,5), rnorm(20, 20, 5)),
                   y = c(rnorm(20), rnorm(20,10,5), rnorm(20, 20, 5)))
cluster_comparison <- compare_clusters(data, k = 2:6)
print(cluster_comparison)
}
