% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_clust_helper.R
\name{.assign_clusters}
\alias{.assign_clusters}
\title{Assign Clusters helper function}
\usage{
.assign_clusters(data, centroids, distance_function)
}
\arguments{
\item{distance_matrix}{a matrix of distances}

\item{centroids_indices}{matrix of centroids}
}
\value{
vector of cluster assignments
}
\description{
Assign Clusters helper function
}
\keyword{internal}
