% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_to_centroid_helper.R
\name{.distance_to_centroid}
\alias{.distance_to_centroid}
\title{Distance to Centroids}
\usage{
.distance_to_centroid(point, centroid, metric)
}
\arguments{
\item{point}{points from matrix}

\item{centroid}{cetroid matrix}

\item{metric}{type of distance}
}
\description{
Distance to Centroids
}
