% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dunn_index_helper.R
\name{.dunn_index}
\alias{.dunn_index}
\title{Dunn Index Helper Function}
\usage{
.dunn_index(clusters, metric)
}
\arguments{
\item{clusters}{a kmeans object output from using calculate_clusters() in this package}

\item{metric}{a distance measure}
}
\description{
Dunn Index Helper Function
}
\keyword{internal}
