% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gower_distance_helper.R
\name{.gower_distance}
\alias{.gower_distance}
\title{Calculate Gower Distances}
\usage{
.gower_distance(point, centroid)
}
\arguments{
\item{point}{these are points}

\item{centroid}{centroid points}
}
\value{
numeric distance values
}
\description{
Calculate Gower Distances
}
\keyword{internal}
