% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lloyds_algorithm_helper.R
\name{.lloyds_algorithm}
\alias{.lloyds_algorithm}
\title{Lloyd's Algorithm}
\usage{
.lloyds_algorithm(data, k, distance_metric, max_iters = 100, tol = 1e-04)
}
\arguments{
\item{data}{distance matrix}

\item{k}{num of clusters}

\item{max_iters}{max iterations}

\item{tol}{tolerance}

\item{distance_func}{distance metric}
}
\value{
vector of cluster assignments
}
\description{
Lloyd's Algorithm
}
\keyword{internal}
