% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_distances_helper.R
\name{.point_distances_helper}
\alias{.point_distances_helper}
\title{Distance Between Points}
\usage{
.point_distances_helper(metric, data)
}
\arguments{
\item{data}{a data matrix}

\item{dist}{distance metric}
}
\description{
Distance Between Points
}
\keyword{internal}
