% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalc_centroids_helper.R
\name{.recalculate_centroids}
\alias{.recalculate_centroids}
\title{Recalculate Centroids Helper Function}
\usage{
.recalculate_centroids(data, cluster_assignments, k)
}
\arguments{
\item{data}{a dataframe}

\item{cluster_assignments}{which cluster belongs to}

\item{k}{num clusters}
}
\value{
matrix of cetroids
}
\description{
Recalculate Centroids Helper Function
}
\keyword{internal}
