% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_helper.R
\name{summary.kmeans_custom}
\alias{summary.kmeans_custom}
\title{Summary Method for Cluster Values}
\usage{
\method{summary}{kmeans_custom}(object, ...)
}
\arguments{
\item{object}{custom kmeans object}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary Method for Cluster Values
}
\keyword{internal}
