% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate Clusters}
\usage{
calculate_clusters(df, k, distance, max.iter = 100, tol = 1e-04, scale = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{k}{an integer number of clusters}

\item{distance}{A distance measure}

\item{max.iter}{Set maximum iterations for convergence}

\item{tol}{tolerance}

\item{scale}{scale data y/n}
}
\value{
CHANGE THIS

matrix of cetroids
}
\description{
Calculate Clusters
}
\keyword{internal}
\keyword{sum(1:10)}
