% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_clusters.R
\name{calculate_clusters}
\alias{calculate_clusters}
\title{Calculate Clusters}
\usage{
calculate_clusters(df, k, distance, max.iter = 100, tol = 1e-04, scale = TRUE)
}
\arguments{
\item{df}{A dataframe}

\item{k}{an integer number of clusters}

\item{distance}{A distance measure}

\item{max.iter}{Set maximum iterations for convergence}

\item{tol}{tolerance}

\item{scale}{scale data y/n}
}
\value{
A list containing the following components:
\item{cluster}{A vector of integers indicating the cluster assignment for each observation.}
\item{centers}{A matrix of cluster centers.}
\item{totss}{The total sum of squares.}
\item{withinss}{The within-cluster sum of squares.}
\item{tot.withinss}{The total within-cluster sum of squares.}
\item{betweenss}{The between-cluster sum of squares.}
\item{size}{The size of each cluster.}
\item{iterations}{The number of iterations for convergence.}
\item{metric}{The distance metric used.}
\item{data}{The input data with cluster assignments appended.}
}
\description{
Calculate Clusters
}
\examples{
data <- mtcars
cluster_values <- calculate_clusters(data,3,"euclidean")
print(cluster_values)


}
