% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_clusters.R
\name{compare_clusters}
\alias{compare_clusters}
\title{Title}
\usage{
compare_clusters(df, k, max.iter = 100, tol = 1e-04, scale = TRUE)
}
\arguments{
\item{df}{a dataframe}

\item{k}{a vector of desired clusters > 1}

\item{max.iter}{maximum number of iterations to convergence}

\item{tol}{distance tollerance}

\item{scale}{TRUE = scale date; FALSE = do no scale data}
}
\value{
Dataframe comparing clusters
}
\description{
Title
}
