% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_clust_helper.R
\name{.assign_clusters}
\alias{.assign_clusters}
\title{Assign Clusters helper function}
\usage{
.assign_clusters(distance_matrix, centroids_indices)
}
\arguments{
\item{data}{dataframe}

\item{centroids}{matrix of centroids}

\item{distance_func}{distance function}
}
\value{
vector of cluster assignments
}
\description{
Assign Clusters helper function
}
\keyword{internal}
