% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lloyds_algorithm_helper.R
\name{.lloyds_algorithm}
\alias{.lloyds_algorithm}
\title{Lloyd's Algorithm}
\usage{
.lloyds_algorithm(data, k, distance_func, max_iters = max_iters, tol = 1e-04)
}
\arguments{
\item{data}{data frame}

\item{k}{num of clusters}

\item{distance_func}{distance metric}

\item{max_iters}{max iterations}

\item{tol}{tolerance}
}
\value{
vector of cluster assignments
}
\description{
Lloyd's Algorithm
}
\keyword{internal}
