% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics_function_helper.R
\name{.metrics_function}
\alias{.metrics_function}
\title{Select Metrics helper}
\usage{
.metrics_function(metric, data)
}
\arguments{
\item{data}{dataframe}

\item{distance}{desired distance (as string)}
}
\value{
calculated distances
}
\description{
Select Metrics helper
}
\keyword{internal}
